pyproject.toml
[project]
name = "nllb-finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.1.1",
    "evaluate>=0.4.6",
    "sacrebleu>=2.5.1",
    "sentencepiece>=0.2.1",
    "torchcodec>=0.7.0",
    "transformers[torch]>=4.56.1",
    "wandb>=0.22.0",
]
